FT-07 винесла функцію в окремий файл
