+++ src/index.js
@@ -1,15 +1,1 @@
-import { FetchMovies } from "./js/fetch";
-
-const fetchMovies = new FetchMovies();
-
-async function loadPopularMovies() {
-
-try {
-    const items = await fetchMovies.fetchPopularMovies();
-        console.log(items)
-    } catch (error) {
-        console.log(error)
-    }    
-}
-
-loadPopularMovies()
+import loadPopularMovies from "./js/popular-movies-fn"
